Split harness server background module
